orders/views.py
from django.shortcuts import render
from rest_framework import viewsets

from .models import Order
# Create your views here.
from .serializers import OrderModelSerializer
from rest_framework.permissions import IsAuthenticated

class OrderViewSet(viewsets.ModelViewSet):
    queryset = Order.objects.all()
    serializer_class = OrderModelSerializer

    def get_permissions(self):
        permission_classes = [IsAuthenticated]
        return [permission() for permission in permission_classes]
